get ui and base styles working in core
