--- apps/web/components/Forms/index.ts
@@ -1,4 +1,0 @@
-export { Form, FORM_ERROR } from "./Form";
-export { TextField } from "./TextField";
-export {RadioGroupField} from "./RadioGroupField";
-export { SelectField } from "./SelectField";
