+++ apps/web/package.json
@@ -15,40 +15,25 @@
   "dependencies": {
     "@dir/core": "workspace:*",
     "@dir/db": "workspace:*",
+    "@dir/ui": "workspace:*",
     "@hookform/error-message": "^2.0.1",
     "@hookform/resolvers": "^3.1.0",
-    "@radix-ui/react-dialog": "^1.0.4",
-    "@radix-ui/react-popover": "^1.0.6",
-    "@radix-ui/react-radio-group": "^1.1.3",
-    "@radix-ui/react-select": "^1.2.2",
-    "@radix-ui/react-switch": "^1.0.3",
-    "@radix-ui/react-tabs": "^1.0.4",
     "@t3-oss/env-nextjs": "^0.3.1",
-    "@tanstack/react-table": "^8.8.5",
-    "class-variance-authority": "^0.6.0",
-    "clsx": "^1.2.1",
-    "cmdk": "^0.2.0",
-    "lucide-react": "^0.263.1",
-    "next": "^14.0.2",
+    "next": "^14.0.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.43.9",
     "server-only": "^0.0.1",
-    "tailwind-merge": "^1.12.0",
     "zod": "^3.20.2"
   },
   "devDependencies": {
     "@types/node": "20.4.4",
     "@types/react": "18.2.16",
     "@types/react-dom": "18.2.7",
-    "autoprefixer": "^10.4.14",
     "dotenv-cli": "^6.0.0",
     "eslint": "^8.30.0",
     "eslint-config-next": "^13.1.1",
     "eslint-plugin-tailwindcss": "^3.11.0",
-    "postcss": "^8.4.23",
-    "tailwindcss": "^3.3.1",
-    "tailwindcss-animate": "^1.0.5",
     "typescript": "5.1.6"
   }
 }
