+++ packages/core/src/features/admin/actions.ts
@@ -1,5 +1,4 @@
 "use server";
-import 'server-only'
 
 import { revalidatePath } from 'next/cache'
 import { prisma } from "@dir/db";
