+++ packages/core/src/features/admin/screens/actions/actions.ts
@@ -1,5 +1,5 @@
 "use server"
-import "server-only"
+
 import { prisma } from "@dir/db";
 import {z} from 'zod'
