+++ packages/core/src/features/admin/screens/categories/actions.ts
@@ -1,5 +1,5 @@
 "use server"
-import "server-only"
+
 import { prisma, type Category } from "@dir/db";
 import { z } from "zod";
