+++ packages/core/src/features/auth/actions.ts
@@ -1,5 +1,5 @@
 "use server";
-import 'server-only'
+
 import {
   PasswordHandler,
   generateToken,
