+++ packages/core/src/index.ts
@@ -3,10 +3,12 @@ import createApiEndpoint from "./lib/createApiEndpoint";
 
 import { authInit } from "./lib/auth";
 import { PageInit, LayoutInit, ApiRouteInit } from "./router";
-import "../dist/output.css";
+
 import type { Tag, Post, User, Item, Category } from "@dir/db";
 import type { PostSchema } from "@/features/posts/schemas";
 
+import "@dir/ui/dist/index.css"
+import "../dist/output.css";
 
 import authDriver from "./authDriver";
 export { PageInit, LayoutInit } from "./router";
