+++ packages/ui/src/components/Skeleton.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { cn } from "@/lib/utils";
+import { cn } from "../utils";
 
 function Skeleton({
   className,
