+++ packages/ui/tsup.config.ts
@@ -1,0 +1,8 @@
+import {defineConfig} from 'tsup'
+
+export default defineConfig({
+    entry: ["src/index.tsx"],
+    outDir: 'dist',
+    dts: true,
+    clean: true
+})
